+++ packages/api/src/models/builders/SearchQueryBuilder.ts
@@ -1,4 +1,4 @@
-import { ParamType, QueryParamType, buildQueryString } from '../helpers'
+import { ParamType } from '../helpers'
 
 /* eslint-disable import/prefer-default-export */
 interface RangeType {
@@ -6,15 +6,27 @@ interface RangeType {
 	max?: string | number | Date
 }
 
+class QueryParam {
+	type: string
+	key: string
+	value: ParamType
+
+	constructor(type: string, key: string, value: ParamType) {
+		this.type = type
+		this.key = key
+		this.value = value
+	}
+}
+
 /**
  * Builder to build AxiosRequestConfig 'search' params
  * for accounting endpoints
  */
 export class SearchQueryBuilder {
-	queryParams: QueryParamType
+	queryParams: QueryParam[]
 
 	constructor() {
-		this.queryParams = {}
+		this.queryParams = [] as QueryParam[]
 	}
 
 	private static handleDate(date: Date): string {
@@ -25,36 +37,52 @@ export class SearchQueryBuilder {
 	}
 
 	in(key: string, values: ParamType[]): SearchQueryBuilder {
-		this.queryParams = { ...this.queryParams, [`search[${key}][]`]: values }
+		this.queryParams.push(new QueryParam('in', key, values))
 		return this
 	}
 
 	equals(key: string, value: ParamType): SearchQueryBuilder {
-		this.queryParams = { ...this.queryParams, [`search[${key}]`]: value }
+		this.queryParams.push(new QueryParam('equals', key, value))
 		return this
 	}
 
 	like(key: string, value: ParamType): SearchQueryBuilder {
-		this.queryParams = { ...this.queryParams, [`search[${key}]`]: value }
+		this.queryParams.push(new QueryParam('like', key, value))
 		return this
 	}
 
 	between(key: string, { min, max }: RangeType): SearchQueryBuilder {
 		if (min) {
 			const value = min instanceof Date ? SearchQueryBuilder.handleDate(min) : min
-			this.queryParams = { ...this.queryParams, [`search[${key}_min]`]: value }
+			this.queryParams.push(new QueryParam('between', `${key}_min`, value))
 		}
 		if (max) {
 			const value = max instanceof Date ? SearchQueryBuilder.handleDate(max) : max
-			this.queryParams = {
-				...this.queryParams,
-				[`search[${key}_max]`]: value,
-			}
+			this.queryParams.push(new QueryParam('between', `${key}_max`, value))
 		}
 		return this
 	}
 
-	build(): string {
-		return buildQueryString(this.queryParams)
+	build(resourceType?: string): string {
+		let isAccountingLike = false
+		if (!resourceType || ['AccountingResource', 'EventsResource'].includes(resourceType)) {
+			isAccountingLike = true
+		}
+		let queryString = ''
+		this.queryParams.forEach((param) => {
+			const type = param.type
+			const key = param.key
+			const value = param.value
+			if (type === 'like' || type === 'between' || (type === 'equals' && isAccountingLike)) {
+				queryString = queryString.concat(`&search[${key}]=${value}`)
+			} else if (type === 'in' && value instanceof Array) {
+				for (let i = 0; i < value.length; i += 1) {
+					queryString = queryString.concat(`&search[${key}][]=${value[i]}`)
+				}
+			} else {
+				queryString = queryString.concat(`&${key}=${value}`)
+			}
+		})
+		return queryString.substr(1)
 	}
 }